Refactor: Extract unresolved/resolved out The logic for filtering unresolved/resolved case has been moved into it's own test (partially) and file.
